llm_friendly_code/requirement.txt
flask>=2.3.2
flask_cors>=3.0.10
openai>=0.27.6
langchain>=0.0.188
bs4>=0.0.1
websocket-client>=1.6.1
simple-websocket>=0.10.1
flask_socketio>=5.3.4
eventlet>=0.33.3

